+++ onecodex/utils.py
@@ -64,9 +64,9 @@ OPTION_HELP = {
     ),
     "prompt": (
         "Manually prompt about automatic paired file interleaving, ONT file merging, and "
-        "multi-lane file concatenation. When prompting, the mate of a paired end file may "
-        "be picked up from the same directory even if it was not passed on the command "
-        "line; it is marked with a * in the prompt. Setting --no-prompt will allow running "
+        "multi-lane file concatenation. When prompting, the rest of a sample's files may "
+        "be picked up from the same directory even if they were not passed on the command "
+        "line; they are marked with a * in the prompt. Setting --no-prompt will allow running "
         "without any user intervention, e.g. in a script, and never uploads a file that "
         "was not passed on the command line."
     ),